// app/(auth)/screens/(employees)/user_account/types.ts
  export interface EmployerDetailsDisplayProps {
    id: string;
        name: string;
        email: string;
        phone: string;
        homeAddress: string;
        dateOfBirth: string;
        Language: string;
        identityVerificationStatus: string;
        backgroundCheckStatus: string;
        description: string;
        rating: number;
        assurance: number;
        urlToImage: any;
        imageStatus: string;
        IDCardStatus: string;
        PassportStatus: string;
        otherCertificateStatus: string;
  }
// ...
  export interface DisplayPersonalInfoProps {
    // key?: string;
    item: any;
    isVerified: boolean;
}
// ...
  export interface UserProps {
    id: string;
    username: string;
    first_name: string;
    last_name: string;
    email: string;
    phone: string;
    home_address: string;
    date_of_birth: Date;
    language: string;
    role: string;
    is_verified: boolean;
    identity_verification: string;
    background_check: string;
  }